Fixed the not-working WYSIWYG-Editor in the custom template.
The template was including jQuery 1.5 while Drupal already was including jQuery 1.4. Drupal also includes a plugin that adds the 'once'-method to the jQuery-Object. This however was overridden again by the jQuery 1.5-import. Also fixed a simple html-mistake (opening h2 but closing h1). Had to make bgstretcher use the JQuery-method because some earlier script seems to disable the $-method.